change app to new version
